Add useContext to share state
App.jsx now serves as single source of truth. State can be read and
update successfully anywhere in the directory.

Side note: Had to debug an issue where the program was crashing when
I tried to update state within a grandchild. I was calling an updater
function in the top level of my functional component, which basically
meant that it was trying to update state while rendering, which it did
not like and ended up causing infinite loops.