Formik sends user data to root state
Formik integration in ProjectForm.tsx now validates user input and
then sends it to App state via the function sendProjectNameToState.

Data is not yet structured how I want it to be but at least I know
formik and yup are functioning and that I can pass state back and
forth.