Add Task Class
The Task class successfully makes Task objects that contain the
necessary data that will be needed for the project later on.

The function called sendTaskToProject must be modified in the future,
but can update nested objects in state as intended. The rest will be
updated when the next form is constructed.