Add useEffect hook to ProjectList to genderate project components
useEffect hook in ProjectList component now successfully generates
a component for every project in the allProjects global state.
When the state is updated, this component will now automatically
add ProjectMenuItem components for every project name in the
allProjects state.

To make this code work, this functionality had to be wrapped in a
useEffect hook, because otherwise map would cause a compile error,
as Object.keys() can only work on objects, but fails if given a
value of null. So basically all this code is written so that there
won't be an error when the app initializes with a value of null bc
no projects are created yet. However, it will 'react' to the change
in global state and update ProjectMenuItem functions dynmamically.