Complete TaskList component
TaskList component now generates TaskMenuItem components based on
how many tasks are assigned to that project. Additionally, the new
task form can add tasks to projects that exist and have them update
the page in real-time.

Like ProjectList, TaskList is written in such a way that it won't
crash if the object it needs to render components from is empty.
Currently it loads the page with mock data if the selected project
has a task property that is empty. Later this will be replaced with
something more user friendly.

Lastly: A bug currently exists that makes it so even when you switch
projects, the old TaskMenuItem components don't go away until you add
a new task to the newly selected project. Instead these components
SHOULD go away as soon as the selectedProject changes.