Create conditional validation schema
EditProjectForm now uses a conditional edit schema that uses a yup
custom test to see if the user is attempting to rename a project a
name that already exists in allProjects.

This was challenging at first because I was trying to bring state
into the yup index.tsx file, which was not possible. However, it
was perfectly possible to define functions that generate
conditional schemas in yup, export them, and then call them in the
relevant component using the state variables that were already in
there.