Add conditional schema to ProjectForm
ProjectForm component now checks existing projects before allowing
the user to submit a name.

Also debugged an issue with the custom schema in which the page
was crashing on initial load because I was trying to use Object.keys
on a null value again. I wrapped this declaration inside an if-
statement checking for null and now it works fine.