Add function updateProjectNameInState()
This function successfully updates the name of a project in the root
state, keeping all of its old tasks. It does so by making a copy
of the tasks, creating a new Project obj with the updated name,
giving it the saved tasks, and THEN writing the new projects to
state.